Documentation: USB permissions and planner times
Fix permissions on Ubuntu and relative instead of absolute default times
in planner

This contains two changes: One documents the fact that in the planner the
default for time is relative and an absolute time needs an @ sign.

The other hopefully explains how to solve permission problems for USB
devices on Ubuntu: I am told, an USB dongle that pretends to be a serial
port by default does not have the required permissions for the standard
user to write. The transfer then fails with an 'obscure' error message.
The transfer works if subsurface is run by root (which obviously is a bad
idea) or with chmod.  Unfortunately, at least Ubuntu 'fixes' permissions
on reboot. So one either has to fiddle with this (seems to be more
complicated) or add the user to the group 'dialout" which gets write
permissions to serial lines.

This is in theory. I don't have access to an Ubuntu installation to see if
all this is indeed the case. That's why I did not send the patch in the
past. But I think this is still better than what we have so far.

Signed-off-by: Robert C. Helling <helling@atdotde.de>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>