Add a make rule (creator-files) that creates files for using Qt Creator
Qt Creator cannot import Subsurface directly because our buildsystem
is not any of the three that it understands (qmake, cmake,
autotools). So, instead, we can create the files Creator uses for
"Other Project" projects.

The files are:
 - subsurface.config:   the #defines from the command line (-D args)
 - subsurface.creator:  an XDG Desktop-style file with Creator settings
 - subsurface.files:    the list of source and header files
 - subsurface.includes: the include paths (-I args)

They are also added to .gitignore, alongside the *.user file that
Creator uses to store per-user settings (editor configuration).

Signed-off-by: Thiago Macieira <thiago@macieira.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>