Implement the get_screen_dpi function
This is sadly known to be incorrect on most systems. But it's as incorrect
as it was on Gtk so should work equally well to get roughly the right
sizes for printing - once we implement printing, that is :-)

Also removed a qDebug that snuck in in commit 6fc4d72079dd ("Enable
defautl_filename in settings").

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>