When editing a manually entered dive, make sure divelist is updated
Editing a manually entered dive allows the user to edit duration, max
depth and mean depth. These values are stored in the first (and only)
divecomputer and the duration, maxdepth and meandepth fields of the dive
are populated in fixup_dive() after the dive was initially parsed.

In order for this to work we need to clear out the existing fields (so
that values can get smaller) and rerun fixup_dive().

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>