Fix yet another selection issue
If we manually deselect dives without changing Gtk's notion of what is
selected and then want to simply recreate the dive list, Gtk re-selects
the unselected dives in the process of clearing out the dive list. How
very strange.

Also, after making changes to the selection we need to make sure that we
update the stats.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>