Fix compiler warnings for maintab.ui
Simply renaming all the elements with name conflict. None of these names
appear to be referenced anywhere so this seems like a rather odd thing to
happen - why doesn't the tool just rename them for me when I open and
close the file in designer? Or at least warn about it?

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>