Merge dive tags when merging dives
.. otherwise the dive tags generally end up cleared when you download a
duplicate dive from another dive computer.

This uses MERGE_NONZERO, which means that if one of the dives has tags
set, we'll prefer those tags.  If both dives have tags set, we take the
tags from the first ("preferred") dive.

We could do a "just or all the bits together" too.  But this way we at
least take a set of tags that are consistent (ie we don't get both
"boat" and "shore" set unless one of the original dives already had that
inconsistency)

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>